- fix wrongly setup mocks
