+++ test/testing/io/read_closer_mock_fixture.go
